requirements.txt
altair==4.2.0
appnope==0.1.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.zoneinfo==0.2.1
blinker==1.5
branca==0.6.0
cachetools==5.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
Fiona==1.8.22
folium==0.14.0
fonttools==4.38.0
geopandas==0.12.2
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.1
ipykernel==6.19.4
ipython==8.7.0
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
munch==2.5.0
nest-asyncio==1.5.6
numpy==1.24.1
packaging==22.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
platformdirs==2.6.1
plotly==5.11.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==10.0.1
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyproj==3.4.1
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pyzmq==24.0.1
requests==2.28.1
rich==12.6.0
seaborn==0.12.1
semver==2.13.0
shapely==2.0.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.16.0
streamlit-folium==0.8.1
tenacity==8.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.8.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
wcwidth==0.2.5
widgetsnbextension==4.0.5
zipp==3.11.0
